requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.8
executing==1.2.0
filelock==3.13.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.41.1
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
multidict==6.0.4
mysqlclient==2.2.0
nest-asyncio==1.5.7
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
regex==2023.8.8
requests==2.31.0
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.2
textblob==0.17.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.25.0
wcwidth==0.2.6
Werkzeug==3.0.0
yarl==1.9.2
